+++ tests/trajectories_from_vec.rs
@@ -11,19 +11,19 @@ fn test_trajectories_from_vec() {
     let mut env_state = Outfit::new("horizon:DE440", ErrorModel::FCCT14).unwrap();
 
     // ---------- Obj 33803 ----------
-    let object_number = "33803";
+    let traj_id = vec![33803];
     let ra_deg = vec![359.7403333333333_f64];
     let dec_deg = vec![-0.5039444444444444_f64];
     let time = vec![43041.93878_f64];
     let observer = env_state.get_observer_from_mpc_code(&"049".to_string()); // Uppsala-Kvistaberg
 
-    let batch = ObservationBatch::from_degrees_owned(&ra_deg, &dec_deg, 0.5, 0.5, &time);
+    let batch = ObservationBatch::from_degrees_owned(&traj_id, &ra_deg, &dec_deg, 0.5, 0.5, &time);
 
     let mut traj_set: TrajectorySet =
-        TrajectorySet::new_from_vec(&mut env_state, object_number, &batch, observer);
+        TrajectorySet::new_from_vec(&mut env_state, &batch, observer).unwrap();
 
     assert_eq!(traj_set.len(), 1);
-    let obs_33803 = traj_set.get(&ObjectNumber::String("33803".into())).unwrap();
+    let obs_33803 = traj_set.get(&ObjectNumber::Int(33803)).unwrap();
     assert_eq!(obs_33803.len(), 1);
 
     assert!((obs_33803[0].time - 43041.93878).abs() < 1e-12);
@@ -39,18 +39,20 @@ fn test_trajectories_from_vec() {
     );
 
     // ---------- Obj 8467 ----------
-    let object_number = "8467";
+    let traj_id = vec![8467];
     let ra_deg = vec![14.62025_f64];
     let dec_deg = vec![9.987777777777778_f64];
     let time = vec![43785.35799_f64];
     let observer: Arc<_> = env_state.get_observer_from_mpc_code(&"675".to_string()); // Palomar Mountain
 
-    let batch = ObservationBatch::from_degrees_owned(&ra_deg, &dec_deg, 0.5, 0.5, &time);
+    let batch = ObservationBatch::from_degrees_owned(&traj_id, &ra_deg, &dec_deg, 0.5, 0.5, &time);
 
-    traj_set.add_from_vec(&mut env_state, object_number, &batch, observer);
+    traj_set
+        .add_from_vec(&mut env_state, &batch, observer)
+        .unwrap();
 
     assert_eq!(traj_set.len(), 2);
-    let obs_8467 = traj_set.get(&ObjectNumber::String("8467".into())).unwrap();
+    let obs_8467 = traj_set.get(&ObjectNumber::Int(8467)).unwrap();
     assert_eq!(obs_8467.len(), 1);
     assert!((obs_8467[0].time - 43785.35799).abs() < 1e-12);